+++ benchmarks-website/store/utils.js
@@ -1,0 +1,3 @@
+export function firstLine(message) {
+  return String(message || "").split("\n")[0] || "";
+}
